batedor_de_ponto_automatizado/config/config.py
from dotenv import load_dotenv
import os

load_dotenv()

load_dotenv(verbose=True)

CREDENCIAIS_AHGORA = {
    "matricula": str(os.getenv("CREDENCIAIS_AHGORA_MATRICULA")),
    "senha": str(os.getenv("CREDENCIAIS_AHGORA_SENHA")),
}